+++ zapit/+zapit/@pointer/stopOptoStim.m
@@ -22,7 +22,7 @@ function stopOptoStim(obj, rampDownInMS)
     end
 
     % Do not try to run if we are presenting a stimulus of a finite length
-    if strcmp(obj.DAQ.hAO.Timing.SampleQuantityMode,'FiniteSamples')
+    if strcmp(obj.DAQ.returnAOSampleQuantityMode,'FiniteSamples')
         return
     end