+++ zapit/+zapit/@stimConfig/stimConfig.m
@@ -5,7 +5,9 @@ classdef stimConfig < handle
     % zapit.stimConfig
     %
     % Purpose
-    % This class handles configuration files used to determine where the laser stim locatins are.
+    % This class handles configuration files used to determine where the laser stim
+    % locations are. Methods in this class generate the waveforms used for stimulus
+    % presentation. This class loads stimulus config YML files and interprets them.
     %
     % Rob Campbell - SWC 2022
 
@@ -22,16 +24,19 @@ classdef stimConfig < handle
 
     properties (Hidden)
         parent  % the zapit.pointer to which this is attached
-        numSamplesPerChannel
+        numSamplesPerChannel % Number of samples per channel to send to the DAQ.
         atlasData    % The atlas data from the loaded .mat file. Shows top-down ARA view in stereotaxic coords
-        edgeSamples  % Samples at which galvoes start to move. (see get.chanSamples) Here for plotChanSamples
         logFileStem = 'zapit_log_' % The stem of the log file name. zapit.pointer will use this to search for the file
+    end
+
+    properties (Hidden, SetAccess=protected)
+        dutyCycle = 0.5 % With one or two stimuli, the beam will always be on half the time.
+                        % It is is with respect to this standard that we set laser powers for all other situations
     end
 
     % read-only properties that are associated with getters
     properties(SetAccess=protected, GetAccess=public)
         chanSamples % The waveforms to be sent to the scanners
-        maxStimPulseDuration
         blankingTime_ms
     end
 
@@ -81,20 +86,6 @@ classdef stimConfig < handle
     % Getters and setters
     methods
 
-        function maxStimPulseDuration = get.maxStimPulseDuration(obj)
-            % Maximum possible stimulus duration given the current blanking time and refresh rate
-            %
-            % zapit.stimConfig.maxStimPulseDuration
-            %
-            % Purpose
-            % There will always be a maximum possible stimulus duration assuming two stimuli
-            % a 50% duty cycle and given frequency. This getter calculates and returns it.
-            % We want the ability to present the stimuli for a shorter time but a higher power.
-            % Therefore we need to know the maximum stimulus duration (a half-cycle).
-            maxStimPulseDuration = (0.5/obj.stimModulationFreqHz)*1E3 - obj.blankingTime_ms;
-
-        end % get.maxStimPulseDuration
-
 
         function blankingTime_ms = get.blankingTime_ms(obj)
             % Return blanking time in ms from the zapit settings
@@ -124,11 +115,11 @@ classdef stimConfig < handle
             % zapit.stimConfig.chanSamples
             %
             % Purpose
-            % The calibratedPoints getter returns the locations of stimulus points in sample space.
-            % The calibratedPointsInVolts converts this to volts. This method uses these voltage
-            % values to generate waveforms that can be played with a clocked NI output task in order
-            % stimulate the sample at the laser power and rep rate defined by properties of this
-            % class.
+            % The calibratedPoints getter returns the locations of stimulus points in
+            % sample space. The calibratedPointsInVolts converts this to volts. This
+            % method uses these voltage values to generate waveforms that can be played
+            % with a clocked NI output task in order stimulate the sample at the laser
+            % power and rep rate defined by properties of this class.
             %
             % Inputs
             % none
@@ -136,16 +127,19 @@ classdef stimConfig < handle
             % Outputs
             % chanSamples
             %
-            % Maja Skretowska - SWC 2021
+            % Original by Maja Skretowska - SWC 2021
             % Updated by Rob Campbell - SWC 2023 to cope with new stimulus structure
 
             % Pull in data from method
 
             if isempty(obj.parent)
+                fprintf(['zapit.pointer not connected to zapit.stimConfig. ', ...
+                    'Not making waveforms.\n'])
                 chanSamples = [];
                 return
             end
 
+            % Cache so it does not get re-generated repeatedly
             calibratedPointsInVolts = obj.calibratedPointsInVolts;
 
             obj.numSamplesPerChannel = obj.parent.DAQ.samplesPerSecond/obj.stimModulationFreqHz;
@@ -155,26 +149,20 @@ classdef stimConfig < handle
             % 2nd dim is channel, 3rd dim is conditions.
             waveforms = zeros(obj.numSamplesPerChannel,4,length(calibratedPointsInVolts));
 
-
-            % Calculate some constants that we will need in multiple places further below.
-
-            % find edges of half cycles (the indexes at which the beam moves or laser changes state)
-            pointsPerTrial = obj.parent.settings.experiment.maxStimPointsPerCondition;
-            obj.edgeSamples = ceil(linspace(1, obj.numSamplesPerChannel, pointsPerTrial+1));
-            sampleInterval = 1/obj.parent.DAQ.samplesPerSecond;
-
-
-
             % The number of samples that correspond to the blanking time. The blanking
             % time is the period during which the beam is off and the scanners slowly
-            % transition from one place to the next. Defined in ms.
+            % transition from one place to the next. This is defined in ms and will be
+            % constant within a trial regardless of how many positions are presented
+            % within a trial.
+            sampleInterval = 1/obj.parent.DAQ.samplesPerSecond;
             blankingSamples = round( (obj.blankingTime_ms*1E-3)/sampleInterval );
 
 
-            % Fill in the matrices for the galvos
+            % STEP ONE -- Fill in the matrices for the galvos
             for ii = 1:length(calibratedPointsInVolts) % Loop over stim conditions
 
                 % If this position is a single point we duplicate it to spoof two points
+                pointsPerTrial = length(calibratedPointsInVolts{ii});
                 if size(calibratedPointsInVolts{ii},1) == 1
                     t_volts = repmat(calibratedPointsInVolts{ii},pointsPerTrial,1);
                 else
@@ -188,53 +176,57 @@ classdef stimConfig < handle
                 % we opt for multiple points.
 
                 % Explicitly extract X and Y scanner voltages from t_volts.
-                % These are column vectors. The first column is the first location and the
-                % second column is the second location.
+                % These are column vectors. The first column is the x scanner voltage
+                % and the second is the y location.
+
                 xVolts = t_volts(:,1)';
                 yVolts = t_volts(:,2)';
 
 
+                % TODO -- update following doc because we now have multiple points.
                 % Make the full waveforms for X and Y. The logic here is that one cycle of
                 % the waveform plays out over numSamplesPerChannel samples. So we want the
-                % beam to be in each of the positions for half the time. We will define
+                % beam to be in each of the two positions for half the time. We will define
                 % this using the number points per trial to make this a bit more explicit
                 % and perhaps more future proof.
-                Y = repmat(yVolts, obj.numSamplesPerChannel/pointsPerTrial, 1);
-                X = repmat(xVolts, obj.numSamplesPerChannel/pointsPerTrial, 1);
+                Y = repmat(yVolts, ceil(obj.numSamplesPerChannel/pointsPerTrial), 1);
+                X = repmat(xVolts, ceil(obj.numSamplesPerChannel/pointsPerTrial), 1);
 
 
                 % The beam will now go to the correct locations but the scanners will
                 % generate a lot of noise because the waveforms have no shaping. We will
                 % therefore swing the scanners slowly from one position to the next over
-                % a period of 1 ms.
+                % a period of roughly 1 ms (exact number is in settings file).
 
-                % This does the ramp at the start of the waveform: it modifies the first column
-                X(1:blankingSamples,1) = linspace(xVolts(2),xVolts(1),blankingSamples);
-                Y(1:blankingSamples,1) = linspace(yVolts(2),yVolts(1),blankingSamples);
+                % This does the ramp at the start of the waveform transitions (even)
+                oldway = false; % WHEN TRUE. WE MAKE WAVEFORMS THE OLD WAY
 
-                % This does the ramp at the middle of the waveform: it modifies the second column
-                X(1:blankingSamples,2) = linspace(xVolts(1),xVolts(2),blankingSamples);
-                Y(1:blankingSamples,2) = linspace(yVolts(1),yVolts(2),blankingSamples);
+                % The following does not produce results identical to the above, but
+                % it's very close. Within a sample.
+                kernel = ones(blankingSamples,1)/blankingSamples;
+                X = X(:);
+                Xsmooth = conv(circshift(X,blankingSamples*2),kernel,'valid');
+                Xsmooth(end:end+blankingSamples-1) = Xsmooth(end);
+                Xsmooth = circshift(Xsmooth,blankingSamples*-1);
+                % Note that the following line (also for Y) will truncate the waveform
+                % slightly. This happens when the number of stimulus locations does not
+                % produce a whole number when divided by the number of stimuli in the
+                % buffer. See lines ~208 and 209 where X and Y are defined. This is
+                % not an easy thing to fix. If we lose a sample or two here and there
+                % from one stimulus it shouldn't matter or be noticeable.
+                waveforms(:,1,ii) = Xsmooth(1:size(waveforms,1));
+                Y = Y(:);
+                Ysmooth = conv(circshift(Y,blankingSamples*2),kernel,'valid');
+                Ysmooth(end:end+blankingSamples-1) = Ysmooth(end);
+                Ysmooth = circshift(Ysmooth,blankingSamples*-1);
+                waveforms(:,2,ii) = Ysmooth(1:size(waveforms,1));
 
-                % Turn the two columns into a row vector and add it into the waveforms array.
-                % Here the first column is the X scan waveform and the second is the Y waveform.
-                waveforms(:,1,ii) = X(:);
-                waveforms(:,2,ii) = Y(:);
-
-                t_mW = obj.laserPowerFromTrial(ii);
-                laserControlVoltage = obj.parent.laser_mW_to_control(t_mW);
-                waveforms(:,3,ii) = ones(1,obj.numSamplesPerChannel) * laserControlVoltage;
-
-                % The masking light
-                waveforms(:,4,ii) = ones(1,obj.numSamplesPerChannel) * 5; % 5V TTL
             end % for ii
 
 
             %%
             % Handling masking for periods beam is moving and one vs two locations
 
-            % Blank the beam and masking light during periods when the beam is moving
-            blankingMask = ones(obj.numSamplesPerChannel,1);
 
 
             % These two lines define variables that allow us to tweak the onset and offset
@@ -247,50 +239,40 @@ classdef stimConfig < handle
             blankOnsetShift_samples = round((blankOnsetShift_ms*1E-3)/sampleInterval);
             blankOffsetShift_samples = round((blankOffsetShift_ms*1E-3)/sampleInterval);
 
-            for ii=1:(blankingSamples+blankOnsetShift_samples+blankOffsetShift_samples)
-                blankingMask(obj.edgeSamples(1:end-1)+(ii-1))=0;
-            end
 
-            waveforms(:,3:4,:) = bsxfun(@times, waveforms(:,3:4,:), blankingMask);
+            % STEP TWO -- Fill in the matrices for the laser power signal
+            waveforms(:,3,:) = 1; % We will fill in with laser power control voltage in sendSamples
+            waveforms(:,4,:) = 5; % 5v TTL
+            for ii = 1:length(calibratedPointsInVolts) % Loop over stim conditions
 
+                pointsPerTrial = length(calibratedPointsInVolts{ii});
+                edgeSamples = ceil(linspace(1, obj.numSamplesPerChannel, pointsPerTrial+1));
 
-            %%
-            % Handle case where we ask for a shorter duration stimulus at higher laser power
-            for ii=1:length(obj.stimLocations)
-
-                if ~isfield(obj.stimLocations(ii).Attributes,'stimPulseDuration_ms')
-                    continue
+                blankingMask = ones(obj.numSamplesPerChannel,1);
+                for kk=1:(blankingSamples+blankOnsetShift_samples+blankOffsetShift_samples)
+                    blankingMask(edgeSamples(1:end-1)+(kk-1))=0;
                 end
 
-                stimDuration = obj.stimLocations(ii).Attributes.stimPulseDuration_ms;
+                waveforms(:,3:4,ii) = waveforms(:,3:4,ii) .* blankingMask;
 
-                if stimDuration < obj.maxStimPulseDuration
-                    % Find the first sample after the beam turns on
-                    digWaveform = waveforms(:,4,ii);
+                % If this has one stim condition we must turn off the beam
+                if size(calibratedPointsInVolts{ii},1) == 1
+                    edgesToZero = edgeSamples(2:2:end)+1;
+                    distanceBetweenEdges = median(diff(edgeSamples));
 
-                    fs = find(diff(digWaveform)>0)+1;
-
-                    % Find the last sample before the beam turns off
-                    fe = find(diff(digWaveform)<0);
-                    fe(end+1) = length(digWaveform);
-
-
-                    % We want the beam on for this long
-                    durationOfStimInSamples = stimDuration*1E-3/sampleInterval;
-
-                    % So we need to blank to zero the following points:
-                    if length(fe)>length(fs)
-                        fs = [1;fs];
+                    for kk=1:length(edgesToZero)
+                        s = edgesToZero(kk);
+                        e = s+distanceBetweenEdges;
+                        if e > obj.numSamplesPerChannel
+                            e = obj.numSamplesPerChannel;
+                        end
+                        waveforms(s:e,3,ii) = 0; % The laser AO line
                     end
-
-                    blankTimes(:,1) = fs+durationOfStimInSamples;
-                    blankTimes(:,2) = fe;
-
-                    % Do it!
-                    waveforms(blankTimes(1,1):blankTimes(1,2),3,ii)=0;
-                    waveforms(blankTimes(2,1):blankTimes(2,2),3,ii)=0;
                 end
+
             end
+
+
 
             %%
             % Handle case where we want an ephys waveform
@@ -304,38 +286,14 @@ classdef stimConfig < handle
             end
 
 
-
-
-
-            % Finally, we loop through and turn off laser on the even cycles when it's a
-            % single position trial. This ensures the beam flashes at the correct rep rate
-            % even though it is stationary.
-
-            % TODO -- I'm sure this can be vectorised (Or do above by making an by 2 array
-            % with second column being zeros).
-            edgesToZero = obj.edgeSamples(2:2:end);
-            distanceBetweenEdges = median(diff(obj.edgeSamples));
-            for ii=1:size(waveforms,3)
-                if size(calibratedPointsInVolts{ii},1) >1
-                    continue
-                end
-
-                for kk=1:length(edgesToZero)
-                    s = edgesToZero(kk);
-                    e = s+distanceBetweenEdges;
-                    if e > size(waveforms,1)
-                        e = size(waveforms,1);
-                    end
-                    waveforms(s:e,3,ii) = 0; % The laser AO line
-                end
-            end
-
-
             % Now we circularly shift the waveforms to deal with the wrapping issue.
-            % TODO: What is the reason this being done???
+            % TODO: What is the reason for doing this?
             waveforms(:,3:4,:) = circshift(waveforms(:,3:4,:),-blankOnsetShift_samples,1);
 
-
+            if size(waveforms,1) ~= obj.numSamplesPerChannel
+                fprintf('Expected waveform size: %d but actual is %d. BAD! REPORT THIS ERROR!', ...
+                    obj.numSamplesPerChannel, size(waveforms,1))
+            end
             % Output
             chanSamples = waveforms;